Test typed-HTTP handlers
The request handler worked greate for handlers that only took a
`*fasthttp.RequestCtx`. But most of our handlers take a project-specific env
object too (a "context").

Create a RequestBuidler[T] type, which largely wraps the existing RequestBuilder
but can be used for these project-specific handlers.

(I don't like the amount of duplication between the two, but it seems like this
is how it has to be done).